# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.RegisterUser:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server
  title: Admin Panel API
  version: "1.0"
paths:
  /auth/login-form:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: success
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "400":
          description: error
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: SignIn
      tags:
      - auth
  /auth/logout-form:
    post:
      consumes:
      - application/json
      description: logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/registration-form:
    post:
      consumes:
      - application/json
      description: create account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          items:
            $ref: '#/definitions/models.RegisterUser'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "303":
          description: error
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: SignUp
      tags:
      - auth
  /auth/sign-in:
    get:
      consumes:
      - application/json
      description: SignIn page
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: ping auth
      tags:
      - auth
  /auth/sign-up:
    get:
      consumes:
      - application/json
      description: SignUp page
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: ping auth
      tags:
      - auth
  /example/helloworld:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /page/admin:
    get:
      consumes:
      - application/json
      description: Admin page
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "400":
          description: Bad Request
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "403":
          description: Forbidden
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "500":
          description: Internal Server Error
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: ping page
      tags:
      - page
  /page/homepage:
    get:
      consumes:
      - application/json
      description: home page
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad Request
          headers:
            Authorization:
              description: Authorization
              type: string
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: ping page
      tags:
      - page
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"